3: Добавит запрос на сохранение записи
Изменит: обращение к календарю(.../calendar), получение определенной записи(.../records?id=1)
