+++ App/Repositories/AbstractRepository.php
@@ -5,6 +5,8 @@ namespace App\Repositories;
 use App\Models\AbstractModel;
 use App\Services\DataBase;
 
+use function PHPSTORM_META\type;
+
 abstract class AbstractRepository implements RepositoryInterface
 {
     protected ?DataBase $dataBase;
@@ -51,7 +53,6 @@ abstract class AbstractRepository implements RepositoryInterface
         $columns = implode(',', $columns);
 
         $sql = "INSERT INTO {$this->tableName} ({$columns}) VALUES ({$paramsValue})";
-
         return $this->save($sql, $paramsList);
     }
